clean code: mtp_verify_extra_state.py
