+++ lightllm/models/qwen3_5/infer_struct.py
@@ -8,7 +8,7 @@ class Qwen35InferStateInfo(Qwen2VLInferStateInfo):
 
     def init_some_extra_state(self, model):
         super().init_some_extra_state(model)
-        from lightllm.common.basemodel.mtp_verify_extra_state import init_mtp_verify_extra_state
+        from lightllm.models.qwen3next.infer_struct import init_mtp_verify_extra_state
 
         init_mtp_verify_extra_state(self, model)
         return
